+++ packages/@cdk-cosmos/core/src/components/core-vpc.ts
@@ -5,18 +5,28 @@ import {
   SubnetType,
   GatewayVpcEndpointAwsService,
   InterfaceVpcEndpointAwsService,
+  IVpc,
+  InterfaceVpcEndpoint,
+  InterfaceVpcEndpointOptions,
 } from '@aws-cdk/aws-ec2';
 import { NetworkBuilder } from '@aws-cdk/aws-ec2/lib/network-util';
+
+export interface ICoreVpc extends IVpc {
+  readonly disableEndpoints?: boolean;
+}
 
 export interface CoreVpcProps extends Partial<VpcProps> {
   networkBuilder: NetworkBuilder;
   cidrMask?: number;
   subnetMask?: number;
+  disableEndpoints?: boolean;
 }
 
-export class CoreVpc extends Vpc {
+export class CoreVpc extends Vpc implements ICoreVpc {
+  readonly disableEndpoints: boolean;
+
   constructor(scope: Construct, id: string, props: CoreVpcProps) {
-    const { networkBuilder, cidrMask = 24, subnetMask = 26 } = props;
+    const { networkBuilder, cidrMask = 24, subnetMask = 26, disableEndpoints } = props;
 
     super(scope, id, {
       maxAzs: 2,
@@ -31,47 +41,64 @@ export class CoreVpc extends Vpc {
       cidr: networkBuilder.addSubnet(cidrMask),
     });
 
+    this.disableEndpoints = disableEndpoints ?? false;
+
     this.addGatewayEndpoint('S3Gateway', {
       service: GatewayVpcEndpointAwsService.S3,
       subnets: [{ subnetGroupName: 'App' }],
     });
   }
+
+  static addCommonEndpoints(vpc: ICoreVpc): void {
+    if (!vpc.disableEndpoints) {
+      tryAddInterfaceEndpoint(vpc, 'SsmEndpoint', {
+        service: InterfaceVpcEndpointAwsService.SSM,
+        subnets: { subnetGroupName: 'App' },
+      });
+      tryAddInterfaceEndpoint(vpc, 'SsmMessageEndpoint', {
+        service: InterfaceVpcEndpointAwsService.SSM_MESSAGES,
+        subnets: { subnetGroupName: 'App' },
+      });
+      tryAddInterfaceEndpoint(vpc, 'CloudWatchLogsEndpoint', {
+        service: InterfaceVpcEndpointAwsService.CLOUDWATCH_LOGS,
+        subnets: { subnetGroupName: 'App' },
+      });
+    }
+  }
+
+  static addEcsEndpoints(vpc: ICoreVpc): void {
+    if (!vpc.disableEndpoints) {
+      tryAddInterfaceEndpoint(vpc, 'EcsEndpoint', {
+        service: InterfaceVpcEndpointAwsService.ECS,
+        subnets: { subnetGroupName: 'App' },
+      });
+      tryAddInterfaceEndpoint(vpc, 'EcsAgentEndpoint', {
+        service: InterfaceVpcEndpointAwsService.ECS_AGENT,
+        subnets: { subnetGroupName: 'App' },
+      });
+      tryAddInterfaceEndpoint(vpc, 'EcsTelemetryEndpoint', {
+        service: InterfaceVpcEndpointAwsService.ECS_TELEMETRY,
+        subnets: { subnetGroupName: 'App' },
+      });
+      tryAddInterfaceEndpoint(vpc, 'EcrEndpoint', {
+        service: InterfaceVpcEndpointAwsService.ECR,
+        subnets: { subnetGroupName: 'App' },
+      });
+      tryAddInterfaceEndpoint(vpc, 'EcrDockerEndpoint', {
+        service: InterfaceVpcEndpointAwsService.ECR_DOCKER,
+        subnets: { subnetGroupName: 'App' },
+      });
+    }
+  }
 }
 
-export function addCommonEndpoints(vpc: Vpc): void {
-  vpc.addInterfaceEndpoint('SsmEndpoint', {
-    service: InterfaceVpcEndpointAwsService.SSM,
-    subnets: { subnetGroupName: 'App' },
-  });
-  vpc.addInterfaceEndpoint('SsmMessageEndpoint', {
-    service: InterfaceVpcEndpointAwsService.SSM_MESSAGES,
-    subnets: { subnetGroupName: 'App' },
-  });
-  vpc.addInterfaceEndpoint('CloudWatchLogsEndpoint', {
-    service: InterfaceVpcEndpointAwsService.CLOUDWATCH_LOGS,
-    subnets: { subnetGroupName: 'App' },
-  });
-}
-
-export function addEcsEndpoints(vpc: Vpc): void {
-  vpc.addInterfaceEndpoint('EcsEndpoint', {
-    service: InterfaceVpcEndpointAwsService.ECS,
-    subnets: { subnetGroupName: 'App' },
-  });
-  vpc.addInterfaceEndpoint('EcsAgentEndpoint', {
-    service: InterfaceVpcEndpointAwsService.ECS_AGENT,
-    subnets: { subnetGroupName: 'App' },
-  });
-  vpc.addInterfaceEndpoint('EcsTelemetryEndpoint', {
-    service: InterfaceVpcEndpointAwsService.ECS_TELEMETRY,
-    subnets: { subnetGroupName: 'App' },
-  });
-  vpc.addInterfaceEndpoint('EcrEndpoint', {
-    service: InterfaceVpcEndpointAwsService.ECR,
-    subnets: { subnetGroupName: 'App' },
-  });
-  vpc.addInterfaceEndpoint('EcrDockerEndpoint', {
-    service: InterfaceVpcEndpointAwsService.ECR_DOCKER,
-    subnets: { subnetGroupName: 'App' },
-  });
-}
+const tryAddInterfaceEndpoint = (
+  vpc: IVpc,
+  id: string,
+  options: InterfaceVpcEndpointOptions
+): InterfaceVpcEndpoint | undefined => {
+  if (!vpc.node.tryFindChild(id)) {
+    return vpc.addInterfaceEndpoint(id, options);
+  }
+  return undefined;
+};