+++ packages/@cdk-cosmos/core/src/cosmos/cosmos-core-import.ts
@@ -1,10 +1,9 @@
 import { Construct, Stack, Fn } from '@aws-cdk/core';
 import { IHostedZone } from '@aws-cdk/aws-route53';
 import { IRepository } from '@aws-cdk/aws-codecommit';
-import { IFunction } from '@aws-cdk/aws-lambda';
 import { ICosmosCore } from '../cosmos/cosmos-core-stack';
 import { BaseConstruct, BaseConstructProps } from '../components/base';
-import { RemoteZone, RemoteCodeRepo, RemoteFunction } from '../helpers/remote';
+import { RemoteZone, RemoteCodeRepo } from '../components/remote';
 
 export interface CosmosCoreImportProps extends BaseConstructProps {}
 
@@ -14,7 +13,7 @@ export class CosmosCoreImport extends BaseConstruct implements ICosmosCore {
   readonly cdkRepo: IRepository;
   readonly rootZone: IHostedZone;
   readonly cdkMasterRoleStaticArn: string;
-  readonly crossAccountExportsFn: IFunction;
+  readonly crossAccountExportServiceToken: string;
 
   constructor(scope: Construct, id: string, props?: CosmosCoreImportProps) {
     super(scope, id, {
@@ -26,11 +25,12 @@ export class CosmosCoreImport extends BaseConstruct implements ICosmosCore {
     this.libVersion = Fn.importValue(this.singletonId('LibVersion'));
     this.cdkRepo = RemoteCodeRepo.import(this, 'CdkRepo', this.singletonId('CdkRepo'));
     this.rootZone = RemoteZone.import(this, 'RootZone', this.singletonId('RootZone'));
-    this.crossAccountExportsFn = RemoteFunction.import(
-      this,
-      'CrossAccountExportsFn',
-      this.singletonId('CrossAccountExportsFn')
-    );
+    // this.crossAccountExportsFn = RemoteFunction.import(
+    //   this,
+    //   'CrossAccountExportsFn',
+    //   this.singletonId('CrossAccountExportsFn')
+    // );
+    this.crossAccountExportServiceToken = Fn.importValue(this.singletonId('CrossAccountExportServiceToken'));
     this.cdkMasterRoleStaticArn = `arn:aws:iam::${Stack.of(scope).account}:role/${this.singletonId('CdkMasterRole')}`;
   }
 }