+++ packages/@cdk-cosmos/core/src/helpers/constants.ts
@@ -1,14 +1,12 @@
-export const AWS_ENV = 'AWS_ENV';
-
-export const COSMOS_PARTITION = 'COSMOS_PARTITION';
-export const COSMOS_VERSION = 'COSMOS_VERSION';
-export const COSMOS_NETWORK_BUILDER = 'COSMOS_NETWORK_BUILDER';
+export const CONTEXT_AWS_ENV = 'AWS_ENV';
+export const CONTEXT_COSMOS_PARTITION = 'COSMOS_PARTITION';
+export const CONTEXT_COSMOS_VERSION = 'COSMOS_VERSION';
+export const CONTEXT_COSMOS_NETWORK_BUILDER = 'COSMOS_NETWORK_BUILDER';
+export const CONTEXT_COSMOS_NAMING = 'COSMOS_NAMING';
 
 export const PATTERN = {
-  SINGLETON_COSMOS: '{Partition}{Resource}+',
-  SINGLETON_GALAXY: '{Partition}{Galaxy}{Resource}+',
-  SINGLETON_SOLAR_SYSTEM: '${Partition}{Galaxy}{SolarSystem}{Resource}+',
   STACK: '{Partition}{Cosmos}{Galaxy}?{SolarSystem}?{Version}?{Type}',
   COSMOS: '{Partition}{Cosmos}{Galaxy}?{SolarSystem}?{Resource}*{Version}?',
+  SINGLETON_COSMOS: '${Partition}{Galaxy}?{SolarSystem}?{Resource}+',
   RESOURCE: '{Resource}+{Version}?',
 };