+++ packages/@cdk-cosmos/core/src/helpers/generate-scope-id.ts
@@ -1,5 +1,5 @@
 import { IConstruct } from '@aws-cdk/core';
-import { PATTERN, COSMOS_PARTITION, COSMOS_VERSION } from './constants';
+import { PATTERN, CONTEXT_COSMOS_PARTITION, CONTEXT_COSMOS_VERSION } from './constants';
 import { IKeyValue, nodeId } from './generate-id';
 
 export interface IScope {
@@ -17,21 +17,7 @@ export function generateSingletonId(props: {
   delimiter?: string;
 }): string {
   const { scope, id, type, partition, version, delimiter } = props;
-  let pattern: string;
-  switch (scope.node.type) {
-    case 'Cosmos':
-      pattern = PATTERN.SINGLETON_COSMOS;
-      break;
-    case 'Galaxy':
-      pattern = PATTERN.SINGLETON_GALAXY;
-      break;
-    case 'SolarSystem':
-      pattern = PATTERN.SINGLETON_SOLAR_SYSTEM;
-      break;
-    default:
-      throw new Error(`Singleton Pattern could not be found for ${scope.node.uniqueId}`);
-  }
-  return generateNodeId({ scope, pattern, id, type, partition, version, delimiter });
+  return generateNodeId({ scope, pattern: PATTERN.SINGLETON_COSMOS, id, type, partition, version, delimiter });
 }
 
 export function generateNodeId(props: {
@@ -69,9 +55,9 @@ export const generateScopeId = (props: {
   if (props.scopes) scopes.push(...props.scopes);
 
   const context = getContext(scopes);
-  const partition = props.partition || scope.node.tryGetContext(COSMOS_PARTITION);
+  const partition = props.partition || scope.node.tryGetContext(CONTEXT_COSMOS_PARTITION);
   if (partition) context.push({ key: 'Partition', value: partition });
-  const version = props.version || scope.node.tryGetContext(COSMOS_VERSION);
+  const version = props.version || scope.node.tryGetContext(CONTEXT_COSMOS_VERSION);
   if (version) context.push({ key: 'Version', value: version });
 
   const pattern = props.pattern || getPattern(scopes) || defaultPattern;