+++ packages/@cdk-cosmos/core/src/helpers/utils.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import * as fs from 'fs';
 import * as path from 'path';
 import { Stack, IConstruct } from '@aws-cdk/core';
@@ -8,6 +9,10 @@ export const isCrossAccount = (x: IConstruct, y: IConstruct, includeRegion?: boo
   const diffAccount = stackX.account !== stackY.account;
   const diffRegion = stackX.region !== stackY.region;
   return includeRegion ? diffAccount && diffRegion : diffAccount;
+};
+
+export const isCrossStack = (x: IConstruct, y: IConstruct): boolean => {
+  return Stack.of(x) !== Stack.of(y);
 };
 
 export const getPackageVersion: () => string = () => {