+++ packages/@cdk-cosmos/core/src/solar-system/solar-system-core-stack.ts
@@ -1,5 +1,5 @@
-import { Construct, Duration, Tag } from '@aws-cdk/core';
-import { IVpc, Vpc } from '@aws-cdk/aws-ec2';
+import { Construct, Duration, Tag, IConstruct } from '@aws-cdk/core';
+import { Vpc } from '@aws-cdk/aws-ec2';
 import { NetworkBuilder } from '@aws-cdk/aws-ec2/lib/network-util';
 import { Certificate, DnsValidatedCertificate } from '@aws-cdk/aws-certificatemanager';
 import {
@@ -12,13 +12,15 @@ import {
 import { isCrossAccount } from '../helpers/utils';
 import { BaseStack, BaseStackProps } from '../components/base';
 import { IGalaxyCore } from '../galaxy/galaxy-core-stack';
-import { CoreVpc, CoreVpcProps, addCommonEndpoints } from '../components/core-vpc';
+import { CoreVpc, CoreVpcProps, ICoreVpc } from '../components/core-vpc';
 import { CrossAccountZoneDelegationRecord } from '../components/cross-account';
-import { RemoteVpc, RemoteZone } from '../helpers/remote';
+import { RemoteVpc, RemoteZone } from '../components/remote';
+
+const SOLAR_SYSTEM_CORE_SYMBOL = Symbol.for('@cdk-cosmos/core.CosmosCore');
 
 export interface ISolarSystemCore extends Construct {
   readonly galaxy: IGalaxyCore;
-  readonly vpc: IVpc;
+  readonly vpc: ICoreVpc;
   readonly zone: IPublicHostedZone;
   readonly privateZone: IPrivateHostedZone;
   readonly networkBuilder?: NetworkBuilder;
@@ -27,9 +29,7 @@ export interface ISolarSystemCore extends Construct {
 
 export interface SolarSystemCoreStackProps extends BaseStackProps {
   vpc?: Vpc;
-  vpcProps?: Partial<CoreVpcProps> & {
-    defaultEndpoints?: boolean;
-  };
+  vpcProps?: Partial<CoreVpcProps>;
   zoneProps?: {
     linkZone?: boolean;
     ttl?: Duration;
@@ -51,41 +51,36 @@ export class SolarSystemCoreStack extends BaseStack implements ISolarSystemCore 
       type: 'SolarSystem',
     });
 
-    const { vpc, certificate = false, vpcProps = {}, zoneProps = {} } = props || {};
-    const { defaultEndpoints = true } = vpcProps;
+    Object.defineProperty(this, SOLAR_SYSTEM_CORE_SYMBOL, { value: true });
+
+    const { vpc, certificate, vpcProps = {}, zoneProps = {} } = props || {};
     const { linkZone = true, ttl = Duration.minutes(30) } = zoneProps;
 
     this.galaxy = galaxy;
 
     if (vpc) this.vpc = vpc as Vpc;
     else {
-      if (!this.networkBuilder) {
-        throw new Error(
-          `NetworkBuilder not found, please define cidr range here (SolarSystem: ${id}) or Galaxy or Cosmos.`
-        );
-      }
+      const networkBuilder = this.networkBuilder || vpcProps.networkBuilder;
+      if (!networkBuilder) throw new Error('Network Builder must be provided.');
 
       this.vpc = new CoreVpc(this, 'Vpc', {
         ...vpcProps,
-        networkBuilder: this.networkBuilder,
+        networkBuilder,
       });
     }
 
-    // Only add endpoints if this component owens the Vpc.
-    if (this.vpc.node.scope === this) {
-      if (defaultEndpoints) addCommonEndpoints(this.vpc);
-    }
+    CoreVpc.addCommonEndpoints(this.vpc);
 
     const rootZone = this.galaxy.cosmos.rootZone;
     this.zone = new PublicHostedZone(this, 'Zone', {
-      zoneName: `${id}.${rootZone.zoneName}`.toLowerCase(),
-      comment: `Core Main Zone for ${id} SolarSystem`,
+      zoneName: `${this.node.id}.${rootZone.zoneName}`.toLowerCase(),
+      comment: `Core Main Zone for ${this.node.id} SolarSystem`,
     });
 
     this.privateZone = new PrivateHostedZone(this, 'PrivateZone', {
       vpc: this.vpc,
-      zoneName: `${id}.internal`.toLowerCase(),
-      comment: `Core Main Private Zone for ${id} SolarSystem`,
+      zoneName: `${this.node.id}.internal`.toLowerCase(),
+      comment: `Core Main Private Zone for ${this.node.id} SolarSystem`,
     });
 
     if (certificate) {
@@ -100,16 +95,16 @@ export class SolarSystemCoreStack extends BaseStack implements ISolarSystemCore 
     if (linkZone) {
       if (isCrossAccount(this.zone, rootZone)) {
         new CrossAccountZoneDelegationRecord(this, 'ZoneDelegation', {
-          ttl,
-          comment: `Core Zone Delegation for ${id} SolarSystem.`,
+          ttl: ttl,
+          comment: `Core Zone Delegation for ${this.node.id} SolarSystem.`,
         });
       } else {
         new ZoneDelegationRecord(this, 'ZoneDelegation', {
           zone: rootZone,
           recordName: this.zone.zoneName,
           nameServers: this.zone.hostedZoneNameServers as string[],
-          ttl,
-          comment: `Core Zone Delegation for ${id} SolarSystem.`,
+          ttl: ttl,
+          comment: `Core Zone Delegation for ${this.node.id} SolarSystem.`,
         });
       }
     }
@@ -118,6 +113,20 @@ export class SolarSystemCoreStack extends BaseStack implements ISolarSystemCore 
     RemoteZone.export(this.zone, this.singletonId('Zone'));
     RemoteZone.export(this.privateZone, this.singletonId('PrivateZone'));
 
-    Tag.add(this, 'cosmos:solarsystem', id);
+    Tag.add(this, 'cosmos:solarsystem', this.node.id);
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  static isSolarSystemCore(x: any): x is SolarSystemCoreStack {
+    return typeof x === 'object' && x !== null && SOLAR_SYSTEM_CORE_SYMBOL in x;
+  }
+
+  static of(construct: IConstruct): SolarSystemCoreStack {
+    const scopes = [construct, ...construct.node.scopes];
+    for (const scope of scopes) {
+      if (SolarSystemCoreStack.isSolarSystemCore(scope)) return scope;
+    }
+
+    throw new Error(`No Galaxy Core Stack could be identified for the construct at path ${construct.node.path}`);
   }
 }