+++ packages/@cosmos-building-blocks/service/src/ec2/index.ts
@@ -1,0 +1,6 @@
+export {
+  CloudWatchAgent,
+  CloudWatchAgentProps,
+  CloudWatchMetricBuilder,
+  CloudWatchMetric,
+} from './ec2-cloud-watch-agent';